rename target directories
