+++ libs/wingsdk/.eslintrc.json
@@ -136,12 +136,12 @@
       {
         "zones": [
           {
-            "target": "src/**/!(*.inflight.ts|*.sim.ts|*.test.ts|exports.ts)",
+            "target": "src/**/!(*.inflight.ts|*.sim.ts|*.test.ts)",
             "from": "src/**/*.inflight.ts",
             "message": "Importing inflight code in a preflight module is not allowed. Add \"// eslint-disable-next-line import/no-restricted-paths\" to disable."
           },
           {
-            "target": "src/**/!(*.inflight.ts|*.sim.ts|*.test.ts|exports.ts)",
+            "target": "src/**/!(*.inflight.ts|*.sim.ts|*.test.ts)",
             "from": "src/**/*.sim.ts",
             "message": "Importing simulator code in a preflight module is not allowed. Add \"// eslint-disable-next-line import/no-restricted-paths\" to disable."
           },
