+++ libs/wingsdk/src/index.ts
@@ -1,6 +1,6 @@
-export * as tfaws from "./tf-aws";
+export * as tfaws from "./target-tf-aws";
 export * as core from "./core";
 export * as fs from "./fs";
 export * as cloud from "./cloud";
-export * as sim from "./sim";
+export * as sim from "./target-sim";
 export * as testing from "./testing";
