+++ libs/wingsdk/src/target-sim/bucket.ts
