+++ libs/wingsdk/src/target-sim/function.ts
