+++ libs/wingsdk/src/target-sim/resource.ts
