+++ libs/wingsdk/src/target-sim/schema-resources.ts
