+++ libs/wingsdk/src/target-sim/util.ts
