+++ libs/wingsdk/test/target-sim/reload.test.ts
@@ -1,7 +1,7 @@
 import * as cloud from "../../src/cloud";
 import * as core from "../../src/core";
-import * as sim from "../../src/sim";
-import { IFunctionClient } from "../../src/sim";
+import * as sim from "../../src/target-sim";
+import { IFunctionClient } from "../../src/target-sim";
 import * as testing from "../../src/testing";
 import { mkdtemp } from "../../src/util";
