+++ libs/wingsdk/test/target-tf-aws/bucket.test.ts
@@ -1,5 +1,5 @@
 import * as cloud from "../../src/cloud";
-import * as tfaws from "../../src/tf-aws";
+import * as tfaws from "../../src/target-tf-aws";
 import { mkdtemp } from "../../src/util";
 import { tfResourcesOf, tfSanitize } from "../util";
