+++ libs/wingsdk/test/target-tf-aws/captures.test.ts
@@ -5,7 +5,7 @@ import {
   QueueInflightMethods,
 } from "../../src/cloud";
 import * as core from "../../src/core";
-import * as tfaws from "../../src/tf-aws";
+import * as tfaws from "../../src/target-tf-aws";
 import { mkdtemp } from "../../src/util";
 import { tfResourcesOf, tfSanitize } from "../util";
