+++ libs/wingsdk/test/target-tf-aws/function.inflight.test.ts
@@ -1,7 +1,7 @@
 import { InvokeCommand, LambdaClient } from "@aws-sdk/client-lambda";
 import { fromUtf8 } from "@aws-sdk/util-utf8-node";
 import { mockClient } from "aws-sdk-client-mock";
-import { FunctionClient } from "../../src/tf-aws/function.inflight";
+import { FunctionClient } from "../../src/target-tf-aws/function.inflight";
 
 const lambdaMock = mockClient(LambdaClient);
