+++ libs/wingsdk/test/target-tf-aws/function.test.ts
@@ -1,7 +1,7 @@
 import * as cdktf from "cdktf";
 import * as cloud from "../../src/cloud";
 import * as core from "../../src/core";
-import * as tfaws from "../../src/tf-aws";
+import * as tfaws from "../../src/target-tf-aws";
 import { mkdtemp } from "../../src/util";
 import { tfResourcesOf, tfSanitize } from "../util";
