+++ libs/wingsdk/test/target-tf-aws/logger.test.ts
@@ -2,7 +2,7 @@ import { Polycons } from "@winglang/polycons";
 import * as cdktf from "cdktf";
 import * as cloud from "../../src/cloud";
 import * as core from "../../src/core";
-import * as tfaws from "../../src/tf-aws";
+import * as tfaws from "../../src/target-tf-aws";
 import { tfResourcesOf, tfSanitize } from "../util";
 
 test("inflight function uses a logger", () => {
